when game is built GameBuilder creates GameBoard
by default GameBoard is created with default parameters, specified in
the static class Game.GameSettings. 

create Game.GameSettings. Assure that getGameData in the Game sets the
GameBoard.

unit test of UserService
whenNumberOfPlayersIs2AndPlayersAreHumanAndComputer tests that the the
game was created and the GameBoard for the game was created and returned
as part of the NewGameResponse.gameData